Add gruvbox_no_background option
Add the `gruvbox_no_background` option to use a background colour of
`none`, and change any references to the `bg` colour to the actual `bg0`
colour, as `bg` cannot be used when the vim colour is `none`.

This means the background colour will be set by the terminal - in my
case, solid black. But more importantly, if the terminal changes the
background colour, it will change in vim too. This is necessary for my
tmux setup where non-active panes have a dark grey background and the
active pane has a black background. If vim sets the background colour
explicitly, tmux cannot change the inactive pane to dark grey and I lose
my visual indicator of the active pane.